requirements.txt
PyCryptodome==3.23.0
schedule==1.2.2
